fix a bug that clients of gfarm_agent just hang on FreeBSD-4.X.
The reason is __exit() function in the userland thread library (libc_r)
on FreeBSD-4.X makes all descriptors blocking mode, thus, as soon as
the parent process calls __exit() via gfarm_daemon(), pthread functions
in the child process become broken, because the implementation needs
non-blocking mode, and the descriptor mode is shared between the
parent and the child.
So, we move open_accepting_socket() after gfarm_daemon() to workaround
this problem.
Fortunately, it's rate that open_accepting_socket() fails, thus I hope
this workaround is acceptable.


git-svn-id: https://svn.code.sf.net/p/gfarm/code/gfarm/trunk@1967 4a5e1ece-6a47-475c-b646-731d720b41d0