If --prefix is specified, don't create gfarm.conf in $CONFIG_PREFIX/$sysconfdir, rather, create it just in $CONFIG_PREFIX/etc.
i.e. If sysconfdir=/home/soda/etc, and if --prefix is not specified,
gfarm.conf will be created in /home/soda/etc.  But if --prefix /var/tmp
is specified, now it will be created in /var/tmp/etc. It would be created
in /var/tmp/home/soda/etc before this change.


git-svn-id: https://svn.code.sf.net/p/gfarm/code/gfarm/trunk@2201 4a5e1ece-6a47-475c-b646-731d720b41d0