create new group "gfarmroot", and split existing administration privileges into "gfarmadm" and "gfarmroot" as follows: - "gfarmadm" group can maintain host/user/group information of Gfarm, - "gfarmroot" group can access/modify any data and any metadata    of the Gfarm filesystem.
The privilege to access/modify any data of the Gfarm filesystem had not existed
(because "gfarmadm" had been able to access/modify only metadata),
but "gfarmroot" group members have the privilege from now on.
The purpose of this "gfarmroot" group is
- to backup a Gfarm filesystem.
- to replicate entire contents of a filesystem node.
etc.

Note that the "gfarmroot" group has no member by default,
and normal users shouldn't have the "gfarmroot" privilege usually.
The recommended ways are something like the followings:
- create a dedicated user for the "gfarmroot" privilege,
  and make the user never do non-administrative tasks.
- temporarily assign a user to the privilege, and remove the privilege ASAP.
  e.g.
  1. "gfgroup -m gfarmroot AN-ADMIN-USER" 
  2. let the AN-ADMIN-USER do an administrative task
  3. "gfgroup -m gfarmroot"


git-svn-id: https://svn.code.sf.net/p/gfarm/code/gfarm_v2/trunk@4211 4a5e1ece-6a47-475c-b646-731d720b41d0