fix #109 (gfmd: [1002222] removal_finishedq_enqueue(): insane state 5).
if a GFARM_ERR_BROKEN_PIPE error happens at gfp_xdr_flush(),
the caller of gfs_client_send_request() tries to make the transaction settled,
but the transaction will be also handled by the disconnect_callback routine
of the asynchronous RPC, because the transaction is not removed from the
asynchronous RPC transaction records.
that makes the transaction doubly handled, and causes the problem.

to fix this problem, gfp_xdr_vsend_async_request() internally calls
gfp_xdr_flush() from now on.


git-svn-id: https://svn.code.sf.net/p/gfarm/code/gfarm_v2/trunk@4706 4a5e1ece-6a47-475c-b646-731d720b41d0