implement lower priority job enqueue function to gfmd thread pool, and make the replica remover thread use it, to avoid that a thread which holds the giant_lock sleeps while adding a gfmd-initiated replication request to the back_channel job queue.
note that this is a theoretical problem which was found by code reading.
currently the bottleneck of massive removal is to write the journal file
(found by tatebe-san).

this is workaround against #986 - possible long giant_lock at massive number of file removable,
and adds the metadb_server_remover_queue_length directive to gfmd.conf


git-svn-id: https://svn.code.sf.net/p/gfarm/code/gfarm_v2/branches/2.7@10447 4a5e1ece-6a47-475c-b646-731d720b41d0