merge r10713 from the 2.7 branch:
disable PTHREAD_PRIO_INHERIT 
to fix #1022 - pthread-related weird behavior of gfmd on CentOS 6.6

this problem does not always happen,
but the following behavior is observed by running "make check" multiple times:

 * delay of response, and network_receive_timeout symptom
 * hard hang, and/or 100% cpu consumption of gfmd
  (this may be resolved after rellay long delay)

The following regession tests tend to show this symptom (but not always):

 * "gfwhomai" or "gfgroup -l gfarmadm", called from bin/am_I_gfarmadm
 * lib/libgfarm/gfarm/gfs_pio_create/autoreplica.sh
 * lib/libgfarm/gfarm/gfs_pio_create/autoreplica4.sh
 * lib/libgfarm/gfarm/gfs_pio_create/autoreplica-repattr-over-ncopy.sh
 * gftool/gfmhost/gfmhost.sh

This symptom may happen with binaries with this fix,
if "make check" with the fixed binaries is invoked
just after the problem happens.
After enough break time, this symptom does not happen with the fixed binaries.
So, this seems to be a kernel problem.

PTHREAD_PRIO_INHERIT was added in r10689
as part of fix for #1015 - too long giant_lock is observed when both massive number of replications and file removals happen simultaneously


git-svn-id: https://svn.code.sf.net/p/gfarm/code/gfarm_v2/branches/2.6@10714 4a5e1ece-6a47-475c-b646-731d720b41d0