gfmd: fix race condition about struct gfarm_metadb_server::tflags
struct gfarm_filesystem:mutex is used to protect it, because
lib/libgfarm/gfarm/filesystem.c cannot access struct mdhost::mutex in gfmd.

another option is to add a mutex to struct gfarm_metadb_server,
but it seems inappropriate, because struct gfarm_metadb_server is
a simple value type.

TO-DO:
Creating a new type which holds struct gfarm_metadb_server and a mutex
is the cleanest solution, but it requires too much change.