docker/dist/centos7a: enable PostgreSQL backend
Because otherwise the config-gfarm command in config.sh,
which is invoked via all.sh, will fail.
The config-gfarm command accepts the -X option only
for the PostgreSQL backend, not for the none backend.

BTW, the following note in commit 3092ccaef was wrong:
NOTE: docker/dev/dist/centos7/src is not buildable with this change, yet

docker/dev/dist/centos7/src just works.